+++ packages/apps/backend/src/apis/app/utils/common.module.ts
@@ -1,6 +1,6 @@
 import { Module, Global } from '@nestjs/common';
 
-import { FirebaseAuthModule } from '@/domain/aggregates/firebase-auth/firebase-auth.module';
+import { FirebaseAuthModule } from '@/firebase-auth/firebase-auth.module';
 import { UserQueryService } from '@/domain/aggregates/user/user.query.service';
 import { RepositoryModule } from '@/repository/repository.module';
