+++ packages/apps/backend/src/apis/app/utils/guards/user-auth.guard.ts
@@ -7,7 +7,7 @@ import { CurrentUserData } from '../decorators/current-user.decorator';
 import { ApiError } from '@/apis/utils/api.error';
 import { IS_PUBLIC_KEY } from '@/apis/utils/decorators/public.decorator';
 import { extractTokenFromHttpHeaders } from '@/apis/utils/extract-token-from-http-headers';
-import { FirebaseAuthService } from '@/domain/aggregates/firebase-auth/firebase-auth.service';
+import { FirebaseAuthService } from '@/firebase-auth/firebase-auth.service';
 import { UserQueryService } from '@/domain/aggregates/user/user.query.service';
 
 @Injectable()
