+++ packages/apps/backend/src/domain/usecases/user/user-usecase.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 
 import { DeleteUserService } from './delete-user.service';
 
-import { FirebaseAuthModule } from '@/domain/aggregates/firebase-auth/firebase-auth.module';
+import { FirebaseAuthModule } from '@/firebase-auth/firebase-auth.module';
 import { UserModule } from '@/domain/aggregates/user/user.module';
 
 @Module({
