+++ packages/apps/backend/src/firebase-auth/utils/dto.ts
